Rename Custom Type
